model adjusts
